+++ src/test/java/com/dlb/chess/test/apicarlos/bugs/fixed/PrintApiCarlosVariousBugs.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.apache.logging.log4j.Logger;
-import org.junit.jupiter.api.Test;
 
 import com.dlb.chess.common.NonNullWrapperCommon;
 import com.github.bhlangonijr.chesslib.Board;
@@ -306,37 +305,4 @@ class PrintApiCarlosVariousBugs {
     }
   }
 
-  @SuppressWarnings("static-method")
-  @Test
-  void testSanGenerationNoOverspecificationForPinnedPieceOne() throws Exception {
-    final MoveList moveList = new MoveList();
-    // 1. d4 e6 2. Nc3 Bb4 3. Nf3 Nc6 4. Ng5 h6 5. Ne4
-
-    moveList.add(new Move(Square.D2, Square.D4));
-    moveList.add(new Move(Square.E7, Square.E6));
-    moveList.add(new Move(Square.B1, Square.C3));
-    moveList.add(new Move(Square.F8, Square.B4));
-    moveList.add(new Move(Square.G1, Square.F3));
-    moveList.add(new Move(Square.B8, Square.C6));
-    moveList.add(new Move(Square.F3, Square.G5));
-    moveList.add(new Move(Square.H7, Square.H6));
-    moveList.add(new Move(Square.G5, Square.E4));
-
-    // correct, last move is not Nge4 because the knight on c3 cannot move to e4
-    assertEquals("d4 e6 Nc3 Bb4 Nf3 Nc6 Ng5 h6 Ne4", moveList.toSan().trim());
-  }
-
-  @SuppressWarnings("static-method")
-  @Test
-  void testSanGenerationNoOverspecificationForPinnedPieceTwo() throws Exception {
-    final MoveList moveList = new MoveList("4k3/8/8/8/1b6/2N5/8/4K1N1 w - - 0 1");
-
-    // correct, first move is not Nge2 because the knight on c3 cannot move to e2
-    moveList.add(new Move(Square.G1, Square.E2)); // Ne2
-    assertEquals("Ne2", moveList.toSan().trim());
-
-    // correct, first move is not Nge2 because the knight on c3 cannot move to e2
-    moveList.add(new Move(Square.E8, Square.E7)); // Ne2
-    assertEquals("Ne2 Ke7", moveList.toSan().trim());
-  }
 }
